Move Collection view stubs under v3
The views here will need to be implemented by
talking to the pulp rest api from the views
