--- galaxy_api/api/v2/serializers/__init__.py
@@ -1,4 +1,0 @@
-#from .collection import CollectionSerializer
-#from .collection import VersionSummarySerializer
-#from .collection import VersionDetailSerializer
-#from .collection import CollectionArtifactSerializer
