--- galaxy_api/api/v2/views/__init__.py
@@ -1,12 +1,0 @@
-
-from .api_view import ApiV2RootView
-# from .collection import CollectionDetailView, CollectionListView
-# from .collection_version import VersionDetailView, VersionListView
-
-__all__ = (
-    'ApiV2RootView',
-#    'CollectionDetailView',
-#    'CollectionListView',
-    'VersionDetailView',
-    'VersionListView',
-)
