+++ galaxy_api/api/v3/serializers/__init__.py
@@ -1,3 +1,10 @@
+from .collection import (
+    CollectionSerializer,
+    CollectionArtifactSerializer,
+    CollectionUploadSerializer,
+    VersionSummarySerializer,
+    VersionDetailSerializer,
+)
 from .namespace import NamespaceSerializer
 from .provider_namespace import ProviderNamespaceSerializer
 from .provider_source import ProviderSourceSerializer
