+++ galaxy_api/api/v3/urls.py
@@ -36,7 +36,6 @@ urlpatterns = [
          NamespaceProviderNamespacesList.as_view(),
          name='namespace_provider_namespaces_list'),
 
-
     path('provider_namespaces',
          ProviderNamespaceList.as_view(),
          name='provider_namespace_list'),
@@ -56,4 +55,31 @@ urlpatterns = [
          UserDetail.as_view(),
          name='user_detail'),
 
+    # TODO/FIXME(akl) - these views will be
+    # implemented as a client against pulp API
+    
+    # path('collections/',
+    #      CollectionListView.as_view(),
+    #      name='collection-list'),
+    # path('collections/<int:pk>/',
+    #      CollectionDetailView.as_view(),
+    #      name='collection-detail'),
+    # path('collections/<str:namespace>/<str:name>/',
+    #      CollectionDetailView.as_view(),
+    #      name='collection-detail'),
+
+    # path('collection-versions/<int:version_pk>/',
+    #      VersionDetailView.as_view(),
+    #      name='version-detail'),
+    # path('collections/<str:namespace>/<str:name>/versions/<str:version>/',
+    #      VersionDetailView.as_view(),
+    #      name='version-detail'),
+
+    # # Collection Version list URLs
+    # path('collections/<int:pk>/versions/',
+    #      VersionListView.as_view(),
+    #      name='version-list'),
+    # path('collections/<str:namespace>/<str:name>/versions/',
+    #      VersionListView.as_view(),
+    #      name='version-list'),
 ]
